templ/concretize2.py - replaces custom pattern.
.vimrc - add minimum viable indentation rules for python.

+++ templ/concretize.py
@@ -11,6 +11,9 @@ if not replacement:
 
 for target in filter(os.path.exists, sys.argv[1:]):
     destination = os.path.basename(target).replace('Empty', replacement)
+    if os.path.exists(destination):
+        print("File '{0}' exists. Not generating.".format(destination))
+        continue
     contents = open(target).read().replace('Empty', replacement)
-    print("Generating {0}...".format(destination))
+    print("Generating '{0}'...".format(destination))
     open(destination, 'w').write(contents)
